[updated] - 修改拼团特价中价钱转换的问题

+++ ZiMaoGou/src/com/sensu/android/zimaogou/activity/SpellOrderDetailsActivity.java
@@ -90,7 +90,7 @@ public class SpellOrderDetailsActivity extends BaseActivity implements View.OnCl
         ((TextView) findViewById(R.id.group_buy_price)).setText("¥" + mGroupDetailsResponse.data.price);
         ((TextView) findViewById(R.id.price_market)).setText("自贸购特价¥" + mGroupDetailsResponse.data.price_goods);
         mOldPriceText.setText("¥" + mGroupDetailsResponse.data.price_market);
-        int saveMoney = Integer.parseInt(mGroupDetailsResponse.data.price_market) - Integer.parseInt(mGroupDetailsResponse.data.price);
+        double saveMoney = Double.parseDouble(mGroupDetailsResponse.data.price_market) - Double.parseDouble(mGroupDetailsResponse.data.price);
         ((TextView) findViewById(R.id.save_money)).setText("立省¥" + String.valueOf(saveMoney));
         ((WebView) findViewById(R.id.web_view)).loadUrl(mGroupDetailsResponse.data.description);
